feat: split lifecycle OTLP plugin
